requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
ndg-httpsclient==0.4.4
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
python-jss==1.5.0
requests==2.18.4
six==1.11.0
urllib3==1.22
